requirements.txt
# Dataverse MCP Server - Production Dependencies
# Generated from pyproject.toml

# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0.2

# HTTP client for Dataverse API
requests>=2.32.4

# Model Context Protocol
fastapi-mcp>=0.3.7
fastmcp
mcp>=1.12.0

# Databricks SDK (optional, for user info endpoints)
databricks-sdk>=0.20.0
